+++ builder.bash
@@ -112,9 +112,16 @@ case $1 in
   ;;
 
   docker-push-arm )
-    echo "> Push docker images to docker hub"
+    echo "> Push docker arm image to docker hub"
     echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin &&\
     docker-compose push image-linux-aarch64
+    echo "Push done"
+  ;;
+
+  docker-push-amd64-static )
+    echo "> Push docker amd64 static image to docker hub"
+    echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin &&\
+    docker-compose push image-linux-amd64-static
     echo "Push done"
   ;;
 
@@ -134,6 +141,18 @@ case $1 in
   ./builder.bash docker-push
 
   echo "> deploy done"
+  ;;
+
+  deploy-amd64-static )
+
+  echo "> Amd64 static deploy started , current branch=$CURRENT_BRANCH"
+  ls -lhS build/*
+
+  ./builder.bash build-backend amd64-static
+  ./builder.bash compress-artifacts
+  ./builder.bash docker-push-amd64-static
+
+  echo "> Amd64 Static deploy done"
   ;;
 
   deploy-arm )